requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.10
Pillow==8.0.1
pycodestyle==2.6.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
